set up files, and inserted code.
